refactor(infra): split pg_store_stats/queries/auxiliary under the 300-line cap (issue #407)
Same mixin-split pattern applied to pg_store.py in this branch's first
commit, applied to the three sibling files flagged over the 300-line
§4.1 cap:

pg_store_auxiliary.py (397 lines) deleted, split into six single-
concern modules — its name was itself a §9 grab-bag violation
("auxiliary" names nothing), so the fix is by responsibility, not by
line count:
  - pg_store_checkpoint.py    — ingest-run + session checkpoints
  - pg_store_prospective.py   — trigger-based (prospective) memory CRUD
  - pg_store_procedural.py    — B1 skill/habit CRUD
  - pg_store_archive.py       — schema-mismatch memory archive
  - pg_store_engram.py        — engram slot allocation (Josselyn & Tonegawa 2020)
  - pg_store_cortical_schema.py — cortical "schema" CRUD (Tse 2007) —
    named to avoid colliding with pg_store_ddl.py's unrelated
    database-DDL "schema" vocabulary

pg_store_queries.py (401 lines) split along its own pre-existing
section comments:
  - pg_store_query_stream.py  — keyset-paginated / cursor streaming reads
  - pg_store_co_access.py     — entity co-access / shared-entity JOINs

pg_store_stats.py (406 lines) split along its own pre-existing
section comments:
  - pg_store_consolidation_stage.py — cascade stage transitions (Kandel 2001)
  - pg_store_cls.py           — CLS queries (McClelland 1995) + oscillatory
                                 state + interference detection

pg_store.py's facade gained the 10 new mixin imports/bases (unchanged
public API — same class, same methods). pg_store_host.py's PgStoreHost
contract gained `get_all_memories_for_decay` (cross-mixin: streaming
iter_memories_for_decay calls it). Two small opportunistic cleanups
while relocating (both behavior-preserving): insert_archive and
get_all_memories_with_embeddings now call `self._bytes_to_vector` /
`self._vector_to_bytes` directly instead of a deferred
`from mcp_server.infrastructure.pg_store import PgMemoryStore` —
unneeded now that the methods live in a mixin composed alongside
PgSerializeMixin.

docs/module-inventory.md's infrastructure/ section fully re-catalogued
for the 16 pg_store_*.py modules now in the facade's mixin family (was
"pg_store.py — PostgreSQL + pgvector persistence" as a single line);
file counts re-measured.

Verified: full suite 7417 passed / 0 skipped / 0 failed (both
`test_hook_receipts.py` agent_briefing tests now pass — fixed upstream
by #403, picked up by this branch's rebase), ruff check + format
clean, pyright zero-diagnostic on mcp_server/. Every resulting file
<=300 lines (largest: pg_store_search.py at 280), every method <=40
lines.
