+++ mcp_server/infrastructure/pg_store_host.py
@@ -161,3 +161,5 @@ class PgStoreHost:
         def _insert_memory_on(
             self, conn: psycopg.Connection[DictRow], data: dict[str, Any]
         ) -> int: ...
+
+        def get_all_memories_for_decay(self) -> list[dict[str, Any]]: ...
